requirements.txt
awscli
requests
Pillow
